Add VR propagation test scene.
Summary:
This diff copies the Meridian test scene in the sample project and adds an OVRCameraRig and teleportation control via controllers and hands.

I also made the map have colors (which match the acoustic material colors) because it was hard to see the white teleportation arc against an entirely white scene.

 {F1957608978}

Reviewed By: isaacmosebrook-meta

Differential Revision: D65632524

fbshipit-source-id: a41c228cd02a22743613d76422ed0a4e89c7e018