Fix Unity 6 build jobs using SandcastleTool package
Summary:
Fix the Build job from SandcastleTools for Unity 6000+.
We don't need to set the gradle properties.

For Presence Platform we changed the job to buildUnityWindows since that works, there seems to still have some issues building with buildUnity that needs to be investigated.

We also removed the artifact creation since it's not used.

Reviewed By: mxw-meta

Differential Revision: D86831184

fbshipit-source-id: 25abf202905a8bba14093b2abec08975f90d3495